Initial commit, issue-13
Move logic verbatum from main to function
